requirements.txt
Django==1.5.1
MySQL-python
Markdown
PyYAML
django-social-auth==0.7.7
django-filter
httplib2
oauth2
python-openid
wsgiref
requests
djangorestframework
django-crispy-forms==1.4.0
djrill==0.7.0
django-gravatar==0.1.0
boto>=2.4  
lxml  
django-model-utils                                         
django-storages==1.1.8
django-celery
twython
django-jsonfield==0.9.4
django-cors-headers
django-mobi==0.1.7
djangorestframework-csv
six==1.4
python-instagram
django-bower